+++ src/types.ts
@@ -222,6 +222,19 @@ export type ClientMethod<P extends Procedure<Type, Type, Type>> = ((
     onProgress?: (progress: P["progress"]["inferOut"]) => void,
     requestId?: string,
   ) => CancelablePromise<P["success"]["inferOut"]>;
+  /**
+   * Send the request to specific nodes, or all nodes.
+   * Returns an array of results, one for each node the request was sent to.
+   * Each result is a {@link PromiseSettledResult}, with also an additional property, the node ID of the request
+   */
+  broadcast: (
+    input: P["input"]["inferIn"],
+    onProgress?: (progress: P["progress"]["inferOut"]) => void,
+    /** Number of nodes to send the request to. Leave undefined to send to all nodes */
+    nodes?: number,
+  ) => Promise<
+    Array<PromiseSettledResult<P["success"]["inferOut"]> & { node: string }>
+  >;
 };
 
 /**